requirements.txt
Django==4.2.7
djangorestframework==3.14.0
torch==2.0.1
torchvision==0.15.2
opencv-python==4.8.1.78
numpy==1.24.3
albumentations==1.3.1
tqdm==4.66.1
Pillow==10.0.1
python-decouple==3.8
django-cors-headers==4.3.1
gunicorn==21.2.0